requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3